+++ src/main/java/com/dailyon/orderservice/common/exception/InvalidParamException.java
@@ -1,0 +1,17 @@
+package com.dailyon.orderservice.common.exception;
+
+import org.springframework.http.HttpStatus;
+
+public class InvalidParamException extends CustomException {
+
+  public static final String MESSAGE = "요청한 값이 올바르지 않습니다.";
+
+  public InvalidParamException() {
+    super(MESSAGE);
+  }
+
+  @Override
+  public HttpStatus getStatusCode() {
+    return HttpStatus.BAD_REQUEST;
+  }
+}
