+++ src/main/java/com/dailyon/orderservice/common/exception/advice/ApiControllerAdvice.java
@@ -1,7 +1,7 @@
 package com.dailyon.orderservice.common.exception.advice;
 
 import com.dailyon.orderservice.common.exception.CustomException;
-import com.dailyon.orderservice.common.exception.response.ErrorResponse;
+import com.dailyon.orderservice.common.response.ErrorResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
